fix: import path error, moved info-card to components folder
