+++ components/card/info-card.tsx
